% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Une fonction pour avoir la distribution de l'âge dans un data.frame}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{data.frame d'une commune ou département}
}
\value{
Renvoie un vecteur avec les différents quartiles de l'âge de mon df: 0\%, 25\%, 50\%, 75\%, 100\%
}
\description{
Prend en compte un data frame et renvoie les quartiles d'âge du df
}
