% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints dans un dataframe}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un dataframe représentant une commune ou un département.}
}
\value{
Un entier correspondant au nombre d'adjoints dans le dataframe.
}
\description{
Cette fonction prend en entrée un dataframe et renvoie le nombre d'adjoints présents dans ce dataframe.
Elle s'assure que la colonne 'Libellé de la fonction' existe avant de compter le nombre d'adjoints.
}
