% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{creer_commune}
\alias{creer_commune}
\title{Création d'un objet de classe Commune}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un data frame contenant les informations d'une commune.}
}
\value{
Un data frame avec la classe "Commune".
}
\description{
Vérifie que le data frame ne contient qu'une seule commune avant de lui attribuer la classe "Commune".
}
