% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{creer_departement}
\alias{creer_departement}
\title{Création d'un objet de classe Département}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{Un data frame contenant les informations d'un département.}
}
\value{
Un data frame avec la classe "Département".
}
\description{
Vérifie que le data frame ne contient qu'un seul département avant de lui attribuer la classe "Département".
}
