% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentaires_data.R
\docType{data}
\name{df_Gers}
\alias{df_Gers}
\title{Données sur les élus du département du Gers}
\format{
\subsection{\code{df_Gers}}{

Un data frame avec 16 lignes et 4932 colonnes :
\describe{
\item{Code.du.département}{Le nombre attribué à un département}
\item{Libellé du département}{Le nom du département}
\item{Nom de l'élu}{Le nom de l'élu}
\item{Prénom de l'élu}{Le prénom de l'élu}
\item{Code sexe}{M si Masculin et F si Feminin}
\item{Date de naissance}{La date de naissance au format dd/mm/aaaa}
...
}
}
}
\source{
Données extraites de "elus_data"
}
\usage{
df_Gers
}
\description{
Un sous-ensemble de données sur les élus du Gers, extraites d'un fichier CSV
contenant des informations telles que le nom, l'âge, la profession. Utilisation pour les exemples.
}
\keyword{datasets}
