% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentaires_data.R
\docType{data}
\name{elus_data}
\alias{elus_data}
\title{Données sur les élus}
\format{
\subsection{\code{elus_data}}{

Un data frame avec 16 lignes et 9752 colonnes :
\describe{
\item{Code.du.département}{Le nombre attribué à un département}
\item{Libellé du département}{Le nom du département}
\item{Nom de l'élu}{Le nom de l'élu}
\item{Prénom de l'élu}{Le prénom de l'élu}
\item{Code sexe}{M si Masculin et F si Feminin}
\item{Date de naissance}{La date de naissance au format dd/mm/aaaa}
...
}
}
}
\source{
Données extraites de "elus_sample.csv"
}
\usage{
elus_data
}
\description{
Un sous-ensemble de données sur les élus, extraites d'un fichier CSV
contenant des informations telles que le nom, l'âge, la profession.
}
\keyword{datasets}
