% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Une fonction pour avoir un graphique du nombre des différents codes professionels}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{data.frame d'une commune ou département}
}
\value{
Renvoie un graphique en barre horizontal de la fréquence des différents codes professionels
}
\description{
Prend en compte un data frame, ne garde que les élus et leurs codes pro puis renvoie un graphique en barre horizontal
}
