% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{plot_commune}
\alias{plot_commune}
\title{Graphique des professions des élus d'une commune}
\usage{
plot_commune(df)
}
\arguments{
\item{df}{data.frame d'une commune.}
}
\value{
Un graphique en barres affichant les professions des élus.
}
\description{
Génère un graphique en barres représentant la distribution des professions des élus d'une commune.
}
\examples{
plot_commune(df_Nantes)
}
