% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{plot_departement}
\alias{plot_departement}
\title{Graphique des professions des élus d'un département}
\usage{
plot_departement(df)
}
\arguments{
\item{df}{data.frame d'un département.}
}
\value{
Un graphique en barres affichant les professions des élus.
}
\description{
Génère un graphique en barres représentant la distribution des professions des élus d'un département.
}
\examples{
plot_departement(df_Gers)
}
