% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_semaine5.R
\name{summary_commune}
\alias{summary_commune}
\title{Résumé des informations d'une commune}
\usage{
summary_commune(x)
}
\arguments{
\item{x}{data.frame représentant une commune.}
}
\value{
Aucune valeur retournée. La fonction affiche directement le résumé dans la console.
}
\description{
Cette fonction prend un dataframe représentant une commune et affiche un résumé des informations, y compris :
\itemize{
\item Le nombre d'élus
\item La distribution des âges
\item L'élu le plus âgé
}
}
\examples{
summary_commune(df_Nantes)
}
